server/requirements.txt
flask==1.0.2
flask-cors==3.0.7
pymongo==3.8.0
Flask-Uploads==0.2.1
Werkzeug==2.0.1
numpy==1.17.3
Pillow==8.1.0
matplotlib==3.3.4